Deposit_Savings_Calculator
This program calculates the time it takes for someone to save for a deposit using the UK govt LISA scheme. Based on the user inputs of the value of the property, deposit percentage and the fixed yearly amount into the LISA (where govt adds 25% of what is deposited yearly into it), the program then prints the total amount needed to save and the time it takes to do this from the LISA alone. It then puts in a recommendation on halving the time if one saves x amount into another savings account, calculated based on the initial user inputs.