Soda can OOP
This program creates a soda can object where the user is able to enter its height in cm and radius from the use of setters and getters. The surface area of a cylinder class then passes through the getters as arguments before printing the result in the console displaying two figures past the decimal point using the printf method.